backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
requests==2.31.0
marshmallow==3.20.1
gunicorn==21.2.0
pytest==7.4.3
black==23.9.1
